+++ plugins/core/reporter/discard_reporter.go
@@ -17,7 +17,7 @@
 
 package reporter
 
-import logv3 "skywalking.apache.org/repo/goapi/collect/logging/v3"
+import logv3 "github.com/apache/skywalking-go/protocols/collect/logging/v3"
 
 type discardReporter struct{}
 
@@ -42,5 +42,8 @@ func (r *discardReporter) ConnectionStatus() ConnectionStatus {
 	return ConnectionStatusDisconnect
 }
 func (r *discardReporter) Close() {
+	// do nothing
+}
+func (r *discardReporter) AddProfileTaskManager(p ProfileTaskManager) {
 	// do nothing
 }