+++ plugins/kratosv2/transport/grpc/server_test.go
@@ -36,7 +36,7 @@ import (
 	nativeGRPC "google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 
-	agentv3 "skywalking.apache.org/repo/goapi/collect/language/agent/v3"
+	agentv3 "github.com/apache/skywalking-go/protocols/collect/language/agent/v3"
 )
 
 func init() {