+++ tools/go-agent/config/loader.go
@@ -83,11 +83,18 @@ type Meter struct {
 }
 
 type GRPCReporter struct {
-	BackendService   StringValue     `yaml:"backend_service"`
-	MaxSendQueue     StringValue     `yaml:"max_send_queue"`
-	Authentication   StringValue     `yaml:"authentication"`
-	CDSFetchInterval StringValue     `yaml:"cds_fetch_interval"`
-	TLS              GRPCReporterTLS `yaml:"tls"`
+	BackendService       StringValue       `yaml:"backend_service"`
+	MaxSendQueue         StringValue       `yaml:"max_send_queue"`
+	Authentication       StringValue       `yaml:"authentication"`
+	CDSFetchInterval     StringValue       `yaml:"cds_fetch_interval"`
+	ProfileFetchInterval StringValue       `yaml:"profile_fetch_interval"`
+	TLS                  GRPCReporterTLS   `yaml:"tls"`
+	Pprof                GRPCReporterPprof `yaml:"pprof"`
+}
+
+type GRPCReporterPprof struct {
+	PprofFetchInterval StringValue `yaml:"pprof_fetch_interval"`
+	PprofFilePath      StringValue `yaml:"pprof_file_path"`
 }
 
 type GRPCReporterTLS struct {